+++ internal/pkg/handler/authorization.go
@@ -4,14 +4,18 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"crypto/x509"
+	"time"
 
-	"github.com/dgrijalva/jwt-go/v4"
 	authz "github.com/koverto/authorization/api"
 	"github.com/koverto/authorization/pkg/claims"
 
+	"github.com/dgrijalva/jwt-go/v4"
 	"github.com/koverto/errors"
 	"github.com/koverto/micro"
 	"github.com/koverto/mongo"
+	"github.com/koverto/uuid"
+	"go.mongodb.org/mongo-driver/bson"
+	mmongo "go.mongodb.org/mongo-driver/mongo"
 )
 
 type Authorization struct {
@@ -24,6 +28,11 @@ type Authorization struct {
 type Config struct {
 	MongoUrl   string `json:"mongourl"`
 	PrivateKey string `json:"privatekey"`
+}
+
+type invalidToken struct {
+	TokenID   *uuid.UUID `bson:"_id"`
+	ExpiresAt time.Time
 }
 
 func New(conf *Config, service *micro.Service) (*Authorization, error) {
@@ -52,9 +61,37 @@ func (a *Authorization) Create(ctx context.Context, in *authz.TokenRequest, out 
 }
 
 func (a *Authorization) Validate(ctx context.Context, in *authz.Token, out *authz.TokenResponse) error {
-	return errors.NotImplemented(a.ID)
+	token, claims, err := a.parseToken(in)
+	if err != nil {
+		return err
+	}
+
+	invalidated, err := a.tokenInvalidated(ctx, claims.ID)
+	out.Success = token.Valid && !invalidated
+
+	return err
 }
 
 func (a *Authorization) Invalidate(ctx context.Context, in *authz.Token, out *authz.TokenResponse) error {
 	return errors.NotImplemented(a.ID)
+}
+
+func (a *Authorization) parseToken(in *authz.Token) (*jwt.Token, *claims.Claims, error) {
+	c := &claims.Claims{}
+	token, err := jwt.ParseWithClaims(in.Token, c, jwt.KnownKeyfunc(jwt.SigningMethodES256, a.privateKey))
+	return token, c, err
+}
+
+func (a *Authorization) tokenInvalidated(ctx context.Context, tokenID *uuid.UUID) (bool, error) {
+	var result *invalidToken
+	filter := bson.M{"_id": tokenID}
+
+	collection := a.client.Collection("invalid_tokens")
+	if err := collection.FindOne(ctx, filter).Decode(result); err == mmongo.ErrNoDocuments {
+		return false, nil
+	} else if err != nil {
+		return false, err
+	}
+
+	return true, nil
 }