+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog][], and this project adheres to
 
 - Protobuf API
 - `Authorization.Create`
+- `Authorization.Validate`
 
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html